ingest/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
GitPython
boto3
qdrant-client
sentence-transformers
torch
transformers
networkx
pydantic
requests
faiss-cpu
minio